% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.mult.R
\name{K.mult}
\alias{K.mult}
\title{计算林下多个点位的开敞度}
\usage{
K.mult(a, b)
}
\arguments{
\item{a}{需要计算开敞度点位的(x,y,z)坐标}

\item{b}{林分中的林木坐标与树高}
}
\value{
各个点位的开敞度
}
\description{
计算林下多个点位的开敞度
}
\examples{
a=data.frame(x=c(25,35,40),y=c(25,21,10),size=c(0,1,2))
b=Understory.Light::b
Understory.light::K.mult(a, b)
}
