% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.single.R
\name{K.single}
\alias{K.single}
\title{计算林下单一点位的开敞度}
\usage{
K.single(a, b)
}
\arguments{
\item{a}{需要计算开敞度的林下点位}

\item{b}{林分中的林木坐标与树高}
}
\value{
计算点位的开敞度
}
\description{
计算林下单一点位的开敞度
}
\examples{
a=data.frame(x=25,y=25,size=0)
b=Understory.light::b
Understory.light::K.single(a, b)
}
