% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.Beam_Scale.S.R
\name{MI.Beam_Scale.S}
\alias{MI.Beam_Scale.S}
\title{计算尺度化加权直射荫蔽度最适的校正系数}
\usage{
MI.Beam_Scale.S(minx, maxx, miny, maxy, b, seq, aaix, baix, mi)
}
\arguments{
\item{minx}{绘图区域横坐标的最小值}

\item{maxx}{绘图区域横坐标的最大值}

\item{miny}{绘图区域纵坐标的最小值}

\item{maxy}{绘图区域纵坐标的最大值}

\item{b}{林分内林木的坐标及树高}

\item{seq}{抽样数据的空间分辨率}

\item{aaix}{最小尺度范围}

\item{baix}{最大尺度范围}

\item{mi}{太阳方位角范围参数，pi/mi为与180°太阳方位角的夹角}
}
\value{
尺度化加权直射荫蔽度最适的校正系数
}
\description{
计算尺度化加权直射荫蔽度最适的校正系数
}
\examples{
Understory.light::MI.Beam_Scale.S(minx=50, maxx=100, miny=50, maxy=100, b, seq=5, aaix=6.2, baix=14, mi=1.5)

}
