% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.Beam_Tan.S.R
\name{MI.Beam_Tan.S}
\alias{MI.Beam_Tan.S}
\title{计算加权进界邻体直射荫蔽度最适的校正系数}
\usage{
MI.Beam_Tan.S(minx, maxx, miny, maxy, b, seq, maxtan, mintan, mi)
}
\arguments{
\item{minx}{绘图区域横坐标的最小值}

\item{maxx}{绘图区域横坐标的最大值}

\item{miny}{绘图区域纵坐标的最小值}

\item{maxy}{绘图区域纵坐标的最大值}

\item{b}{林分内林木的坐标及树高}

\item{seq}{抽样数据的空间分辨率}

\item{maxtan}{最大临界值}

\item{mintan}{最小临界值}

\item{mi}{太阳方位角范围参数，pi/mi为与180°太阳方位角的夹角}
}
\value{
加权进界邻体直射荫蔽度最适的校正系数
}
\description{
计算加权进界邻体直射荫蔽度最适的校正系数
}
\examples{
b=Understory.light::b
Understory.light::MI.Beam_Tan.S(minx=50, maxx=100, miny=50, maxy=100, b, seq=5, maxtan=2, mintan=0.5, mi=1.5)
}
