% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.Diffuse_Scale.S.R
\name{MI.Diffuse_Scale.S}
\alias{MI.Diffuse_Scale.S}
\title{计算尺度化加权散射荫蔽度在特定局域尺度下最适的校正系数}
\usage{
MI.Diffuse_Scale.S(minx, maxx, miny, maxy, b, seq, scale)
}
\arguments{
\item{minx}{研究区域横坐标的最小值}

\item{maxx}{研究区域横坐标的最大值}

\item{miny}{研究区域纵坐标的最小值}

\item{maxy}{研究区域纵坐标的最大值}

\item{b}{林分内林木的坐标及树高}

\item{seq}{抽样数据的空间分辨率}

\item{scale}{局域尺度圆的半径}
}
\value{
尺度化加权散射荫蔽度在特定局域尺度下最适的校正系数
}
\description{
计算尺度化加权散射荫蔽度在特定局域尺度下最适的校正系数
}
\examples{
b=Understory.light::b
Understory.light::MI.Diffuse_Scale.S(minx=50, maxx=100, miny=50, maxy=100, b, seq=5, scale=5.4)
}
