% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.Diffuse_Tan.S.R
\name{MI.Diffuse_Tan.S}
\alias{MI.Diffuse_Tan.S}
\title{计算加权进界邻体散射荫蔽度在特定局域临界值下最适的校正系数}
\usage{
MI.Diffuse_Tan.S(minx, maxx, miny, maxy, b, seq, tan)
}
\arguments{
\item{minx}{研究区域横坐标的最小值}

\item{maxx}{研究区域横坐标的最大值}

\item{miny}{研究区域纵坐标的最小值}

\item{maxy}{研究区域纵坐标的最大值}

\item{b}{林分内林木的坐标及树高}

\item{seq}{抽样数据的空间分辨率}

\item{tan}{局域临界值}
}
\value{
加权进界邻体散射荫蔽度在特定局域临界值下最适的校正系数
}
\description{
计算加权进界邻体散射荫蔽度在特定局域临界值下最适的校正系数
}
\examples{
b=Understory.light::b
Understory.light::MI.Diffuse_Tan.S(minx=50, maxx=100, miny=50, maxy=100, b, seq=5, tan=2.8)
}
