% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.K.Krig.R
\name{plot.K.Krig}
\alias{plot.K.Krig}
\title{绘制林下开敞度的kring插值图}
\usage{
\method{plot}{K.Krig}(minx, maxx, miny, maxy, b, seq)
}
\arguments{
\item{minx}{绘图区域横坐标的最小值}

\item{maxx}{绘图区域横坐标的最大值}

\item{miny}{绘图区域纵坐标的最小值}

\item{maxy}{绘图区域纵坐标的最大值}

\item{b}{林分内林木的坐标及树高}

\item{seq}{绘制插值图时抽样点的空间分辨率}
}
\value{
林下开敞度的kring插值图与半变异函数
}
\description{
绘制林下开敞度的kring插值图
}
\examples{
b=Understory.light::b
Understory.light::plot.K.Krig(minx=50,maxx=100,miny=50,maxy=100,b,seq=20)
}
