% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.K.single.R
\name{plot.K.single}
\alias{plot.K.single}
\title{绘制单个研究兴趣点的开敞度示意图}
\usage{
\method{plot}{K.single}(a, b)
}
\arguments{
\item{a}{研究兴趣点的(x,y,z)坐标}

\item{b}{林分内林木的坐标及树高}
}
\value{
单个研究兴趣点的开敞度示意图
}
\description{
绘制单个研究兴趣点的开敞度示意图
}
\examples{
a=data.frame(x=75,y=75,size=0)
b=Understory.light::b
Understory.light::plot.K.single(a,b)
}
