% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Scale.dependence.Diffuse_S.Krig.R
\name{plot.Scale.dependence.Diffuse_S.Krig}
\alias{plot.Scale.dependence.Diffuse_S.Krig}
\title{绘制尺度化加权散射荫蔽度的Kring插值图}
\usage{
\method{plot}{Scale.dependence.Diffuse_S.Krig}(minx, maxx, miny, maxy, b, seq, scale, MI, wei, scale2)
}
\arguments{
\item{minx}{绘图区域横坐标的最小值}

\item{maxx}{绘图区域横坐标的最大值}

\item{miny}{绘图区域纵坐标的最小值}

\item{maxy}{绘图区域纵坐标的最大值}

\item{b}{样地内林木的横坐标与树高}

\item{seq}{抽样数据的空间分辨率}

\item{scale}{局域尺度圆半径}

\item{MI}{校正系数}

\item{wei}{线性校正系数}

\item{scale2}{背景尺度圆的半径，可以省略}
}
\value{
绘制尺度化加权散射荫蔽度的Kring插值图
}
\description{
绘制尺度化加权散射荫蔽度的Kring插值图
}
\examples{
b=Understory.light::b
Understory.light::plot.Scale.dependence.Diffuse_S.Krig(minx=50, maxx=100, miny=50, maxy=100, b=b, seq=20, scale=5.4, MI=10.813,wei=0.5)
Understory.light::plot.Scale.dependence.Diffuse_S.Krig(minx=0, maxx=50, miny=0, maxy=50, b=b, seq=20, scale=5.4, MI=10.813, wei=0.5, scale2=15.5)

}
