% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Tan.dependence.Beam_S.Krig.R
\name{plot.Tan.dependence.Beam_S.Krig}
\alias{plot.Tan.dependence.Beam_S.Krig}
\title{绘制加权进界邻体直射直射荫蔽度的Kring插值图}
\usage{
\method{plot}{Tan.dependence.Beam_S.Krig}(minx, maxx, miny, maxy, b, seq, maxtan, mintan, mi, MI, wei, bandwidth)
}
\arguments{
\item{minx}{绘图区域横坐标最小值}

\item{maxx}{绘图区域横坐标最大值}

\item{miny}{绘图区域纵坐标最小值}

\item{maxy}{绘图区域纵坐标最大值}

\item{b}{样地内林木的横坐标与树高}

\item{seq}{绘制插值图时抽样点的空间分辨率}

\item{maxtan}{最大临界值}

\item{mintan}{最小临界值}

\item{mi}{太阳方位角范围参数，pi/mi为与180°太阳方位角的夹角}

\item{MI}{校正系数}

\item{wei}{线性校正系数}

\item{bandwidth}{林木分布太阳方位角核概率密度函数的带宽}
}
\value{
加权进界邻体直射直射荫蔽度的Kring插值图
}
\description{
绘制加权进界邻体直射直射荫蔽度的Kring插值图
}
\examples{
b=Understory.light::b
Understory.light::plot.Tan.dependence.Beam_S.Krig(minx=50, maxx=100, miny=50, maxy=100, b=b, seq=20, maxtan=2, mintan=0.5, mi=1.5, MI=17.8, wei=0.5, bandwidth=0.5)

}
