% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Tan.dependence.Diffuse_S.Krig.R
\name{plot.Tan.dependence.Diffuse_S.Krig}
\alias{plot.Tan.dependence.Diffuse_S.Krig}
\title{绘制加权进界邻体散射荫蔽度的Kring插值图}
\usage{
\method{plot}{Tan.dependence.Diffuse_S.Krig}(minx, maxx, miny, maxy, b, seq, tan, MI, wei, tan2)
}
\arguments{
\item{minx}{绘图区域横坐标的最小值}

\item{maxx}{绘图区域横坐标的最大值}

\item{miny}{绘图区域纵坐标的最小值}

\item{maxy}{绘图区域纵坐标的最大值}

\item{b}{样地内林木的横坐标与树高}

\item{seq}{抽样数据的空间分辨率}

\item{tan}{局域临界值}

\item{MI}{校正系数}

\item{wei}{线性校正系数}

\item{tan2}{背景临界值，可以省略}
}
\value{
加权进界邻体散射荫蔽度的Kring插值图
}
\description{
绘制加权进界邻体散射荫蔽度的Kring插值图
}
\examples{
b=Understory.light::b
Understory.light::plot.Tan.dependence.Diffuse_S.Krig(minx=50, maxx=100, miny=50, maxy=100, b=b, seq=20, tan=2.8, MI=10.660, wei=0.5)
}
