% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Tan.dependence.Diffuse_S.single.R
\name{plot.Tan.dependence.Diffuse_S.single}
\alias{plot.Tan.dependence.Diffuse_S.single}
\title{绘制单一林下位点的加权进界邻体散射荫蔽度的示意图}
\usage{
\method{plot}{Tan.dependence.Diffuse_S.single}(a, b, tan, MI, wei, tan2)
}
\arguments{
\item{a}{林下研究位点的横坐标、纵坐标与高度}

\item{b}{样地内林木的横坐标与树高}

\item{tan}{局域临界值}

\item{MI}{校正系数}

\item{wei}{线性校正系数}

\item{tan2}{背景临界值，可以省略}
}
\value{
单一林下位点的加权进界邻体散射荫蔽度的示意图
}
\description{
绘制单一林下位点的加权进界邻体散射荫蔽度的示意图
}
\examples{
a=data.frame(x=75,y=75,size=0)
b=Understory.light::b
Understory.light::plot.Tan.dependence.Diffuse_S.single(a=a,b=b,tan=2.8,MI=10.660,wei=0.5)
Understory.light::plot.Tan.dependence.Diffuse_S.single(a=a,b=b,tan=2.8,MI=10.660,wei=0.5,tan2=1)
}
