% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale.dependence.Beam_S.mult.R
\name{Scale.dependence.Beam_S.mult}
\alias{Scale.dependence.Beam_S.mult}
\title{计算林下多个位点的尺度化加权直射荫蔽度}
\usage{
Scale.dependence.Beam_S.mult(a, b, aaix, baix, mi, MI, wei, bandwidth)
}
\arguments{
\item{a}{林下研究位点的横坐标、纵坐标与高度}

\item{b}{样地内林木的横坐标与树高}

\item{aaix}{最小尺度范围}

\item{baix}{最大尺度范围}

\item{mi}{太阳方位角范围参数，pi/mi为与180°太阳方位角的夹角}

\item{MI}{校正系数}

\item{wei}{线性校正系数}

\item{bandwidth}{林木分布太阳方位角核概率密度函数的带宽}
}
\value{
林下多个位点的尺度化加权直射荫蔽度
}
\description{
计算林下多个位点的尺度化加权直射荫蔽度
}
\examples{
a=data.frame(x=c(25,35,40),y=c(25,21,10),size=c(0,1,2))
b=Understory.light::b
Understory.light::Scale.dependence.Beam_S.mult(a=a, b=b, aaix=6.2, baix=14, mi=1.5, MI=11.2, wei=0.5, bandwidth=0.5)
}
