% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale.dependence.Diffuse_S.single.R
\name{Scale.dependence.Diffuse_S.single}
\alias{Scale.dependence.Diffuse_S.single}
\title{计算单一林下位点的尺度化加权散射荫蔽度}
\usage{
Scale.dependence.Diffuse_S.single(a, b, scale, MI, wei, scale2)
}
\arguments{
\item{a}{林下研究位点的横坐标、纵坐标与高度}

\item{b}{样地内林木的横坐标与树高}

\item{scale}{局域尺度圆半径}

\item{MI}{校正系数}

\item{wei}{线性校正系数}

\item{scale2}{背景尺度圆的半径，可以省略}
}
\value{
单一林下位点的尺度化加权散射荫蔽度
}
\description{
计算单一林下位点的尺度化加权散射荫蔽度
}
\examples{
a=data.frame(x=75,y=75,size=0)
b=Understory.light::b
Understory.light::Scale.dependence.Diffuse_S.single(a=a,b=b,scale=5.4,MI=10.813,wei=0.5)
}
