% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tan.dependence.Beam_S.mult.R
\name{Tan.dependence.Beam_S.mult}
\alias{Tan.dependence.Beam_S.mult}
\title{计算林下多个位点的加权进界邻体直射荫蔽度}
\usage{
Tan.dependence.Beam_S.mult(a, b, maxtan, mintan, mi, MI, wei, bandwidth)
}
\arguments{
\item{a}{林下研究位点的横坐标、纵坐标与高度}

\item{b}{样地内林木的横坐标与树高}

\item{maxtan}{最大临界值}

\item{mintan}{最小临界值}

\item{mi}{太阳方位角范围参数，pi/mi为与180°太阳方位角的夹角}

\item{MI}{校正系数}

\item{wei}{线性校正系数}

\item{bandwidth}{林木分布太阳方位角核概率密度函数的带宽}
}
\value{
林下多个位点的加权进界邻体直射荫蔽度
}
\description{
计算林下多个位点的加权进界邻体直射荫蔽度
}
\examples{
a=data.frame(x=c(25,35,40),y=c(25,21,10),size=c(0,1,2))
b=Understory.Light::b
Understory.light::Tan.dependence.Beam_S.mult(a=a, b=b, maxtan=2, mintan=0.5, mi=1.5, MI=17.8, wei=0.5, bandwidth=0.5)
}
