% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tan.dependence.Beam_S.single.R
\name{Tan.dependence.Beam_S.single}
\alias{Tan.dependence.Beam_S.single}
\title{计算单一林下位点的加权进界邻体直射荫蔽度}
\usage{
Tan.dependence.Beam_S.single(a, b, maxtan, mintan, mi, MI, wei, bandwidth)
}
\arguments{
\item{a}{林下研究位点的横坐标、纵坐标与高度}

\item{b}{样地内林木的横坐标与树高}

\item{maxtan}{最大临界值}

\item{mintan}{最小临界值}

\item{mi}{太阳方位角范围参数，pi/mi为与180°太阳方位角的夹角}

\item{MI}{校正系数}

\item{wei}{线性校正系数}

\item{bandwidth}{林木分布太阳方位角核概率密度函数的带宽}
}
\value{
单一林下位点的加权进界邻体直射荫蔽度
}
\description{
计算单一林下位点的加权进界邻体直射荫蔽度
}
\examples{
a=data.frame(x=75,y=75,size=0)
b=Understory.light::b
Understory.light::Tan.dependence.Beam_S.single(a=a,b=b, maxtan=2, mintan=0.5, mi=1.5, MI=17.8, wei=0.5, bandwidth=0.5)

}
